Refactor message app option helpers
