+++ scripts/notifyUsers.ts
@@ -23,11 +23,7 @@ import {
   getJORFSearchLinkPeople
 } from "../utils/JORFSearch.utils.ts";
 import Organisation from "../models/Organisation.ts";
-import { ExternalMessageOptions, sendMessage } from "../entities/Session.ts";
-import { WhatsAppAPI } from "whatsapp-api-js/middleware/express";
-import { ErrorMessages } from "../entities/ErrorMessages.ts";
-import { WHATSAPP_API_VERSION } from "../entities/WhatsAppSession.ts";
-import { SignalCli } from "signal-sdk";
+import { sendMessage } from "../entities/Session.ts";
 import {
   NotificationTask,
   dispatchTasksToMessageApps
@@ -40,23 +36,15 @@ import {
 } from "matrix-bot-sdk";
 */
 import { sendMainMenu } from "../commands/default.ts";
+import {
+  parseEnabledMessageApps,
+  resolveExternalMessageOptions
+} from "../utils/messageAppOptions.ts";
 
 // Number of days to go back: 0 means we just fetch today's info
 const SHIFT_DAYS = 30;
 
-const { ENABLED_APPS } = process.env;
-if (ENABLED_APPS === undefined) throw new Error("ENABLED_APPS env var not set");
-const enabledApps = JSON.parse(ENABLED_APPS) as MessageApp[];
-
-const invalidApps: string[] = [];
-for (const app of enabledApps) {
-  if (!["Telegram", "Matrix", "WhatsApp", "Signal"].includes(app))
-    invalidApps.push(app);
-}
-if (invalidApps.length > 0)
-  throw new Error(
-    `Invalid message app${invalidApps.length > 0 ? "s" : ""}: ${invalidApps.join(", ")}`
-  );
+const enabledApps = parseEnabledMessageApps();
 
 /*
 let matrixClient: MatrixClient | undefined = undefined;
@@ -77,40 +65,7 @@ if (enabledApps.includes("Matrix")) {
 }
 */
 
-let whatsAppAPI: WhatsAppAPI | undefined = undefined;
-if (enabledApps.includes("WhatsApp")) {
-  const { WHATSAPP_USER_TOKEN, WHATSAPP_APP_SECRET, WHATSAPP_VERIFY_TOKEN } =
-    process.env;
-  if (
-    WHATSAPP_USER_TOKEN === undefined ||
-    WHATSAPP_APP_SECRET === undefined ||
-    WHATSAPP_VERIFY_TOKEN === undefined
-  )
-    throw new Error(ErrorMessages.WHATSAPP_ENV_NOT_SET);
-
-  whatsAppAPI = new WhatsAppAPI({
-    token: WHATSAPP_USER_TOKEN,
-    appSecret: WHATSAPP_APP_SECRET,
-    webhookVerifyToken: WHATSAPP_VERIFY_TOKEN,
-    v: WHATSAPP_API_VERSION
-  });
-}
-
-let signalCli: SignalCli | undefined = undefined;
-if (enabledApps.includes("Signal")) {
-  const { SIGNAL_BAT_PATH, SIGNAL_PHONE_NUMBER } = process.env;
-  if (SIGNAL_BAT_PATH === undefined || SIGNAL_PHONE_NUMBER === undefined)
-    throw new Error(ErrorMessages.SIGNAL_ENV_NOT_SET);
-
-  signalCli = new SignalCli(SIGNAL_BAT_PATH, SIGNAL_PHONE_NUMBER);
-  await signalCli.connect();
-}
-
-const messageAppsOptions: ExternalMessageOptions = {
-  //matrixClient,
-  signalCli: signalCli,
-  whatsAppAPI: whatsAppAPI
-};
+const messageAppsOptions = await resolveExternalMessageOptions(enabledApps);
 
 async function getJORFRecordsFromDate(
   startDate: Date
